requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# RAG & LLM
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langchain-chroma==0.2.2
chromadb==0.6.3
tiktoken==0.8.0
openai==1.59.3

# Voice / Vapi
httpx==0.28.1
websockets==14.1
aiohttp==3.11.11

# Data & utilities
numpy==2.2.1
pandas==2.2.3
jinja2==3.1.5
pyyaml==6.0.2
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Logging & monitoring
structlog==24.4.0
